+++ Demos/UEPLyraIntegration/Plugins/UEPLyraBridge/Source/UEPLyraBridge/UEPLyraBridge.Build.cs
@@ -19,6 +19,7 @@ public class UEPLyraBridge : ModuleRules
             new string[]
             {
                 "GameFeatures",
+                "GameplayAbilities",
                 "LyraGame"
             }
         );